(PC-2287) : remove mounted tests from venue container to a separate file because of strange behaviour and complexity, eslint does not appreciate return in test
